Change in JJB to be able send messages to ci message bus
To be able generate jobs, which send messages to ci message bus,
it is necessary to add config file to get_job_as_xml. The config
is created as a temporary and it contains the jenkins master
URL and credentials (if it is required). Credentials are obtained
from environment variables.